+++ src/components/code/_code_block.js
@@ -120,7 +120,7 @@ export class EuiCodeBlockImpl extends Component {
       {
         'euiCodeBlock--transparentBackground': transparentBackground,
         'euiCodeBlock--inline': inline,
-        'euiCodeBlock-isCopyable': isCopyable,
+        'euiCodeBlock--hasControls': isCopyable || overflowHeight,
       },
       className
     );
@@ -130,7 +130,7 @@ export class EuiCodeBlockImpl extends Component {
     const optionalStyles = {};
 
     if (overflowHeight) {
-      optionalStyles.height = overflowHeight;
+      optionalStyles.maxHeight = overflowHeight;
     }
 
     const codeSnippet = (
@@ -255,7 +255,9 @@ export class EuiCodeBlockImpl extends Component {
 
     return (
       <div {...wrapperProps}>
-        <pre className="euiCodeBlock__pre">{codeSnippet}</pre>
+        <pre style={optionalStyles} className="euiCodeBlock__pre">
+          {codeSnippet}
+        </pre>
 
         {/*
           If the below fullScreen code renders, it actually attaches to the body because of